Enum binrw::Endian

pub enum Endian {
    Big,
    Little,
}

Defines the order of bytes in a multi-byte type.

Variants

Big

The most significant byte is stored first.

Little

The least significant byte is stored first.

Implementations

impl Endian

pub const NATIVE: Self = Endian::Little

The target platformโs native endianness.

pub fn from_utf16_bom_bytes(bom: [u8; 2]) -> BinResult<Self>

Converts a byte array containing a UTF-16 byte order mark into an Endian value.

Errors

Returns an error if the input does not contain a byte order mark.

pub fn into_utf16_bom_bytes(self) -> [u8; 2]

Converts an Endian value into an array containing a UTF-16 byte order mark.