Module binrw::helpers

Helper functions for reading and writing data.

Functions

args_iter

Creates a parser that builds a collection using items from the given iterable object as arguments for the parser.

args_iter_with

Creates a parser that uses a given function to build a collection, using items from the given iterable object as arguments for the function.

count

Creates a parser that reads N items into a collection.

count_with

Creates a parser that uses a given function to read N items into a collection.

read_u24

Reads a 24-bit unsigned integer.

until

Creates a parser that reads items into a collection until a condition is met. The terminal item is added to the collection.

until_eof

Creates a parser that reads items into a collection until the end of the input stream.

until_eof_with

Creates a parser that uses a given function to read items into a collection until the end of the input stream.

until_exclusive

Creates a parser that reads items into a collection until a condition is met. The terminal item is discarded.

until_exclusive_with

Creates a parser that uses a given function to read items into a collection until a condition is met. The terminal item is discarded.

until_with

Creates a parser that uses a given function to read items into a collection until a condition is met. The terminal item is added to the collection.

write_u24

Writes a 24-bit unsigned integer.