Module binrw::io

Traits, helpers, and type definitions for core I/O functionality.

By default, this module simply re-exports the parts of std::io that are used by binrw. In no_std environments, a compatible subset API is exposed instead.

Modules

prelude

The I/O Prelude

Structs

BufReader

A wrapper to add buffering to a read stream.

Bytes

An iterator over u8 values of a reader.

Cursor

A Cursor wraps an in-memory buffer and provides it with a Seek implementation.

Error

The error type for I/O operations of the Read, Write, Seek, and associated traits.

NoSeek

A wrapper that provides a limited implementation of Seek for unseekable Read and Write streams.

TakeSeek

Read adapter which limits the bytes read from an underlying reader, with seek support.

Enums

ErrorKind

A list specifying general categories of I/O error.

SeekFrom

Enumeration of possible methods to seek within an I/O object.

Traits

Read

The Read trait allows for reading bytes from a source.

Seek

The Seek trait provides a cursor which can be moved within a stream of bytes.

TakeSeekExt

An extension trait that implements take_seek() for compatible streams.

Write

A trait for objects which are byte-oriented sinks.

Type Aliases

Result

A specialized Result type for I/O operations.