Struct binrw::io::BufReader

pub struct BufReader<T> { /* private fields */ }

A wrapper to add buffering to a read stream.

Unlike std::io::BufReader, this wrapper does not invalidate the read buffer every time a Seek method is called. It also caches the underlying stream position to avoid unnecessary system calls.

Limitations

Reading or seeking the wrapped stream object directly will cause an inconsistency in the internal state of the BufReader. Calling BufReader::seek_invalidate will clear the read buffer and reset the internal state to be consistent with the wrapped stream.

Implementations

impl<T: Read> BufReader<T>

pub fn new(inner: T) -> BufReader<T>

Creates a new BufReader<T> with a default buffer capacity.

pub fn with_capacity(capacity: usize, inner: T) -> BufReader<T>

Creates a new BufReader<T> with the specified buffer capacity.

impl<T> BufReader<T>

pub fn buffer(&self) -> &[u8]

Returns a reference to the internally buffered data.

pub fn capacity(&self) -> usize

Returns the number of bytes the internal buffer can hold at once.

pub fn get_mut(&mut self) -> &mut T

Gets a mutable reference to the underlying reader.

It is inadvisable to directly read from the underlying reader as it will, at the least, break the cached position information.

pub fn get_ref(&self) -> &T

Gets a reference to the underlying reader.

pub fn into_inner(self) -> T

Unwraps this BufReader<T>, returning the underlying reader.

Note that any leftover data in the internal buffer is lost. Therefore, a following read from the underlying reader may lead to data loss.

impl<T: Seek> BufReader<T>

pub fn seek_invalidate(&mut self, pos: SeekFrom) -> Result<u64>

Performs a seek that forces invalidation of the buffer and internal position state.

Errors

Returns an error if seeking fails.