Struct binrw::io::NoSeek

pub struct NoSeek<T> { /* private fields */ }

A wrapper that provides a limited implementation of Seek for unseekable Read and Write streams.

This is useful when reading or writing from unseekable streams where binrw does not actually need to seek to successfully parse or write the data.

Implementations

impl<T> NoSeek<T>

pub fn new(inner: T) -> Self

Creates a new seekable wrapper for the given value.

pub fn get_mut(&mut self) -> &mut T

Gets a mutable reference to the underlying value.

pub fn get_ref(&self) -> &T

Gets a reference to the underlying value.

pub fn into_inner(self) -> T

Consumes this wrapper, returning the underlying value.