Module binrw::prelude

The binrw prelude.

A collection of traits and types you’ll likely need when working with binrw and are unlikely to cause name conflicts.

use binrw::prelude::*;

Re-exports

pub use crate::binrw;

pub use crate::BinRead;

pub use crate::BinReaderExt;

pub use crate::BinResult;

pub use crate::BinWrite;

pub use crate::BinWriterExt;

Attribute Macros

binread

Attribute macro used to generate an impl of the trait BinRead with support for temporary variables.

binrw

Attribute macro used to generate an impl of both BinRead and BinWrite traits with support for temporary variables.

binwrite

Attribute macro used to generate an impl of the trait BinWrite with support for temporary variables.

Derive Macros

BinRead

Derive macro generating an impl of the trait BinRead.

BinWrite

Derive macro generating an impl of the trait BinWrite.