Trait binrw::NamedArgs

pub trait NamedArgs {
    type Builder;

    // Required method
    fn builder() -> Self::Builder;
}

The NamedArgs trait allows named arguments objects to be constructed using a builder that checks for correctness at compile time.

See #[derive(NamedArgs)] for information on deriving custom named arguments types.

Required Associated Types

type Builder

The builder type for this type.

Required Methods

fn builder() -> Self::Builder

Creates a new builder for this type.

Object Safety

This trait is not object safe.